docs: ➕ add PredicateType usage 📄
