+++ README.md
@@ -164,8 +164,8 @@ If you use any [functions](/lib/functions) you should add `ts-essentials` to you
 - [`PredicateFunction`](/lib/predicate-function.ts) - Matches type constraint for type guard, meaning first argument is
   used in return type and return type is
   [type predicate](https://www.typescriptlang.org/docs/handbook/2/narrowing.html#using-type-predicates)
-- [`PredicateType<Type>`](/lib/predicate-type.ts) - Constructs a type which equals to narrowed type in predicate
-  function `Type`
+- [`PredicateType<Type>`](/lib/predicate-type) - Constructs a type which equals to narrowed type in predicate function
+  `Type`
 
 ### Utility functions
