+++ lib/predicate-type/index.ts
@@ -1,0 +1,8 @@
+import { PredicateFunction } from "../predicate-function";
+
+export type PredicateType<Type extends PredicateFunction> = Type extends (
+  target: any,
+  ...rest: any[]
+) => target is infer NarrowedType
+  ? NarrowedType
+  : never;
